+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/user/byemail/GetUserByEmailHandler.java
@@ -1,8 +1,8 @@
-package io.ciphernance.identity.application.query.user.getuserbyemail;
+package io.ciphernance.identity.application.query.user.byemail;
 
 import io.ciphernance.identity.application.exception.user.UserNotFoundException;
 import io.ciphernance.identity.application.mediator.QueryHandler;
-import io.ciphernance.identity.application.query.user.getuserprofile.GetUserProfileResponse;
+import io.ciphernance.identity.application.query.user.profile.GetUserProfileResponse;
 import io.ciphernance.identity.domain.model.User;
 import io.ciphernance.identity.domain.port.UserRepositoryPort;
 import io.ciphernance.identity.domain.vo.Email;
