+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/user/profile/GetUserProfileHandler.java
@@ -1,4 +1,4 @@
-package io.ciphernance.identity.application.query.user.getuserprofile;
+package io.ciphernance.identity.application.query.user.profile;
 
 import io.ciphernance.identity.application.exception.user.UserNotFoundException;
 import io.ciphernance.identity.application.mediator.QueryHandler;
