+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/user/profile/GetUserProfileQuery.java
@@ -1,4 +1,4 @@
-package io.ciphernance.identity.application.query.user.getuserprofile;
+package io.ciphernance.identity.application.query.user.profile;
 
 import io.ciphernance.identity.application.mediator.Query;
